Separating the ThreeJS related classes into different modules for better testing; Separating the end-to-end testing script into different scripts for open mode and headless mode, like the separation between the unit testing scripts (Visual Testing not working properly in headless mode)
